+++ package.json
@@ -18,7 +18,9 @@
   "homepage": "https://github.com/9659tig/P-1#readme",
   "dependencies": {
     "body-parser": "^1.19.1",
+    "ejs": "^3.1.6",
     "express": "^4.17.2",
-    "mysql": "^2.18.1"
+    "mysql": "^2.18.1",
+    "mysql2": "^2.3.3"
   }
 }
