--- public/page/promboard.js
@@ -1,37 +1,0 @@
-export default function Homepage(board) {
-    board.innerHTML  =
-        `
-        <section class="board">
-            <div class="top">
-                <div>홍보게시판</div>
-                <div class="assign">등록하기</div>
-            </div>
-
-
-            <div class="boardlist">
-                <ul>
-                    <li><a href="#a"><p class="date">01/04</p><div>아주비기너샘플아주비기너샘플아주비기너샘플아주비기너샘플아주비기너샘플아주비기너샘플아주비기너샘플아주비기너샘플아주비기너샘플</div><p class="name">익명</p></a></li>
-                    <li><a href="#a"><p class="date">01/04</p><div>아주비기너샘플아주비기너샘플아주비기너샘플아주비기너샘플아주비기너샘플아주비기너샘플아주비기너샘플아주비기너샘플아주비기너샘플</div><p class="name">익명</p></a></li>
-                    <li><a href="#a"><p class="date">01/04</p><div>아주비기너샘플아주비기너샘플아주비기너샘플아주비기너샘플아주비기너샘플아주비기너샘플아주비기너샘플아주비기너샘플아주비기너샘플</div><p class="name">익명</p></a></li>
-                    <li><a href="#a"><p class="date">01/04</p><div>아주비기너샘플아주비기너샘플아주비기너샘플아주비기너샘플아주비기너샘플아주비기너샘플아주비기너샘플아주비기너샘플아주비기너샘플</div><p class="name">익명</p></a></li>
-                    <li><a href="#a"><p class="date">01/04</p><div>아주비기너샘플아주비기너샘플아주비기너샘플아주비기너샘플아주비기너샘플아주비기너샘플아주비기너샘플아주비기너샘플아주비기너샘플</div><p class="name">익명</p></a></li>
-                    <li><a href="#a"><p class="date">01/04</p><div>아주비기너샘플아주비기너샘플아주비기너샘플아주비기너샘플아주비기너샘플아주비기너샘플아주비기너샘플아주비기너샘플아주비기너샘플</div><p class="name">익명</p></a></li>
-                </ul>
-            </div>
-        </section>
-
-        <footer class="pageindex">
-            <ul>
-                <li><a href="#a">1</a></li>
-                <li><a href="#a">2</a></li>
-                <li><a href="#a">3</a></li>
-                <li><a href="#a">4</a></li>
-                <li><a href="#a">5</a></li>
-                <li><a href="#a">6</a></li>
-                <li><a href="#a">7</a></li>
-                <li><a href="#a">8</a></li>
-                <li><a href="#a">9</a></li>
-            </ul>
-        </footer>
-        `;
-}
