+++ public/view/join.js
