+++ public/view/login.js
@@ -1,0 +1,12 @@
+export default function Login(loginPage) {
+    loginPage.innerHTML = `
+    <section class="loginPage">
+        <div>
+            <input type="text" placeholder="ID를 입력하세요.">
+            <button class="join">회원가입</button>
+            <button class="realLogin">로그인</button>
+        </div>
+    </section>
+
+    `;
+}
