+++ public/view/write.js
